+++ packages/react-drop-indicator/examples/constellation/box/box-indent.tsx
@@ -5,8 +5,7 @@
 
 import { useEffect, useState } from 'react';
 
-// eslint-disable-next-line @atlaskit/ui-styling-standard/use-compiled -- Ignored via go/DSP-18766
-import { css, jsx } from '@emotion/react';
+import { css, jsx } from '@compiled/react';
 
 import { Stack, Text } from '@atlaskit/primitives';
 import { token } from '@atlaskit/tokens';